Edited error throwing

+++ src/ui/components/slider/slider.h
@@ -11,7 +11,7 @@
 class Slider
 {
 private:
-  double min, max, value = 0.5;
+  float min, max, value = 0.5;
 
   int getKnobX();
   int getKnobY();
@@ -20,7 +20,7 @@ private:
 public:
   int w = 300, h = 20,
       x, y;
-  double *assign;
+  float *assign;
   TEXT comment;
   SDL_Color knobTextColor = {0, 0, 0, 255};
   SDL_Color knobColor = {128, 128, 128, 255};
@@ -28,14 +28,14 @@ public:
 
   bool isHeld = false;
 
-  double getValue();
-  void setValue(double _value);
+  float getValue();
+  void setValue(float _value);
 
   void Draw(SDL_Renderer *renderer);
 
   void Update(Inputs &inputs);
 
-  Slider(double _min, double _max, int _x, int _y, double *assign = nullptr);
+  Slider(float _min, float _max, int _x, int _y, float *assign = nullptr);
 };
 
 #endif
